Consolidate chat protocol types
